requirements.txt
openai
streamlit
python-dotenv
langchain
langchain-openai
black
isort
llama-index
psycopg2-binary
boto3
fastapi
uvicorn
pydantic